mqtt_subscriber/requirements.txt
paho-mqtt
python-dotenv
pymongo
requests>=2.0